Add entitties.py
