--- src/experiments/base_experiment.py
@@ -1,10 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Experiment(ABC):
-    def __init__(self, name):
-        self.name = name
-
-    @abstractmethod
-    def create(self):
-        pass
